fix: resolve circular import by extracting PhaseResult to separate module
- Extract PhaseResult to src/agentmux/workflow/phase_result.py
- Import PhaseResult from base in event_router.py
- Update all handlers to use PhaseResult from base
